[RENAME] 변경된 상위클래스에 맞게 improt경로 수정
